Replace strings with TextSendable objects from Texts to make messages consistent.

+++ src/main/java/me/blayyke/vgmgroups/command/groups/child/CommandChildGroupDisband.java
@@ -2,6 +2,7 @@ package me.blayyke.vgmgroups.command.groups.child;
 
 import com.google.common.collect.Lists;
 import me.blayyke.vgmgroups.Group;
+import me.blayyke.vgmgroups.Texts;
 import me.blayyke.vgmgroups.VGMGroups;
 import me.blayyke.vgmgroups.command.Command;
 import me.blayyke.vgmgroups.manager.GroupManager;
@@ -9,14 +10,10 @@ import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
-import org.spongepowered.api.command.args.CommandElement;
-import org.spongepowered.api.command.args.GenericArguments;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.text.Text;
-import org.spongepowered.api.text.format.TextColors;
 
 import javax.annotation.Nonnull;
-import java.util.Optional;
 
 public class CommandChildGroupDisband extends Command {
     public CommandChildGroupDisband(VGMGroups plugin) {
@@ -34,8 +31,13 @@ public class CommandChildGroupDisband extends Command {
         Player player = playersOnly(src);
         Group group = requireGroup(player);
 
-        player.sendMessage(Text.of(TextColors.RED, "Your group has been deleted"));
+        if (!group.getRank(player.getUniqueId()).getRank().isOwner()) {
+            Texts.OWNER_ONLY.send(player);
+            return CommandResult.empty();
+        }
+
         GroupManager.getInstance().deleteGroup(group);
+        Texts.GROUP_DISBANDED.send(player);
 
         return CommandResult.success();
     }
